- Update Stock Watcher demo; add periodic refresh and user input functionality. - Support quote-delimited values in CSVSerializer. - Fix minor validation error in ListView and TableView. - Use long values for timeouts in setInterval() and setTimeout() instead of int. - Fix minor repaint bug in LinkButtonSkin.
git-svn-id: https://svn.apache.org/repos/asf/incubator/pivot/trunk@753699 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/pivot/wtk/ApplicationContext.java
@@ -855,7 +855,7 @@ public abstract class ApplicationContext {
      *
      * @return An integer ID that can be used to cancel execution of the task.
      */
-    public static int setInterval(Runnable runnable, int period) {
+    public static int setInterval(Runnable runnable, long period) {
         int intervalID = nextTimerTaskID++;
 
         IntervalTask intervalTask = new IntervalTask(runnable);
@@ -886,7 +886,7 @@ public abstract class ApplicationContext {
      *
      * @return
      */
-    public static int setTimeout(Runnable runnable, int timeout) {
+    public static int setTimeout(Runnable runnable, long timeout) {
         int timeoutID = nextTimerTaskID++;
 
         TimeoutTask timeoutTask = new TimeoutTask(runnable, timeoutID);
